take the initialValue from the formik
