+++ website/frontend/src/components/pages/Send/children/MessageForm/children/PayloadType/index.ts
@@ -1,0 +1,1 @@
+export * from "./PayloadType";
