--- website/frontend/src/components/pages/Send/children/MessageForm/children/SwitchableField/index.ts
@@ -1,1 +1,0 @@
-export * from "./SwitchableField"
